+++ spec/pg_cron_rails/statements_spec.rb
@@ -1,9 +1,0 @@
-require "spec_helper"
-
-module PgCronRails
-  RSpec.describe 'Statements' do
-    it "must not fail if pg_cron extension is not enabled" do
-      expect { ActiveRecord::Base.connection.schedule_pg_cron_job("dummy_job") }.to_not raise_error
-    end
-  end
-end